Adding filtered dataset
