+++ submissions/2024-11-27-OpenFF-NAGL2-Training-Optimization-Dataset-v4.0/dataset.json.bz2
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b37a02829ee5ab933fd587963587eb7601212ebc634a8fa988b75580d94e049e
+size 36272532
